доделал второй экран
