--- UIKitDZ/UIKitDZ/NotificationsViewController.swift
@@ -1,15 +1,0 @@
-// NotificationsViewController.swift
-// Copyright © RoadMap. All rights reserved.
-
-import UIKit
-
-/// экран уведомлений
-class NotificationsViewController: UIViewController {
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-        view.backgroundColor = .white
-
-        print("ViewDidLoadMessages")
-    }
-}
